# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Point of Sales API",
        "description": "API for Point of Sales System",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/transactions/my": {
            "get": {
                "tags": [
                    "Transactions (Cashier)"
                ],
                "summary": "Get cashier transaction list",
                "description": "Internal transaction list (not receipt)",
                "operationId": "b8742bd1a2e5844ab15885ed01163c50",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter by transaction_code (LIKE)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter start date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter end date (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "day | week | month | year (Legacy)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Custom start date (Legacy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Custom end date (Legacy)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Success",
                                    "data": {
                                        "current_page": 1,
                                        "per_page": 10,
                                        "total": 3,
                                        "data": [
                                            {
                                                "transaction_id": 3,
                                                "transaction_code": "TRX-20260131212326-XGGJ",
                                                "transaction_date": "2026-01-31 21:23:26",
                                                "total_amount": "46000.00",
                                                "payment_amount": "50000.00",
                                                "change_amount": "4000.00",
                                                "status": "completed"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions/my/{id}": {
            "get": {
                "tags": [
                    "Transactions (Cashier)"
                ],
                "summary": "Get cashier transaction detail",
                "description": "Show detail of a transaction created by the authenticated cashier",
                "operationId": "5b398989284b4a4543b9b4915d8342e8",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "application/json"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "Success",
                                    "data": {
                                        "transaction_code": "TRX-20260131212326-XGGJ",
                                        "transaction_date": "2026-01-31 21:23:26",
                                        "status": "completed",
                                        "total_amount": "46000.00",
                                        "payment_amount": "50000.00",
                                        "change_amount": "4000.00",
                                        "items": [
                                            {
                                                "product_id": 5,
                                                "product_name": "Nasi Kuning Spesial",
                                                "quantity": 2,
                                                "price": "18000.00",
                                                "subtotal": "36000.00"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Transaction not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "message": "Transaction not found or access denied"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/transactions": {
            "post": {
                "tags": [
                    "Transactions (Cashier)"
                ],
                "summary": "Create new transaction (Stateless)",
                "operationId": "3a793d45761159f5c7f545d95d5c2e00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items",
                                    "paid_amount"
                                ],
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "example": "cash"
                                    },
                                    "paid_amount": {
                                        "type": "number",
                                        "example": 50000
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "qty": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaction created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}